requirements.txt
certifi==2020.6.20
cloudpickle==1.6.0
cycler==0.10.0
future==0.18.2
gym==0.17.3
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.1
np==1.0.2
numpy==1.19.1
pandas==1.1.1
Pillow==7.2.0
pyglet==1.5.0
pymdptoolbox==4.0b3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
